renamed availability file
